# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Capture API Simulator API",
    "description": "Deterministic traffic-forensics simulator. Browsers cannot call it directly (no CORS): proxy it from your own server. Optional fields are absent rather than null; times are ISO-8601 UTC with milliseconds; session ids are uint64 decimal strings.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/hunts": {
      "get": {
        "tags": [
          "hunts"
        ],
        "summary": "List your saved hunts",
        "description": "Only the caller's own hunts, newest first. Hunts are never shared.",
        "operationId": "listHunts",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HuntList"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "hunts"
        ],
        "summary": "Save a hunt",
        "description": "Answers 201 with `Location` and `ETag: \"v1\"`.",
        "operationId": "createHunt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HuntCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hunt"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Semantic validation error",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorBody"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Createhunt"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/hunts/{hunt_id}": {
      "get": {
        "tags": [
          "hunts"
        ],
        "summary": "Read one saved hunt",
        "description": "404 `hunt_not_found` for an unknown id and for another user's hunt.",
        "operationId": "getHunt",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "hunt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hunt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hunt"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "hunts"
        ],
        "summary": "Update a saved hunt",
        "description": "JSON merge-patch: omitted keys stay as they are and `description: null` removes the description. `Content-Type` may be `application/merge-patch+json` or plain `application/json`. `If-Match` is REQUIRED: without it the answer is 428 `precondition_required`, with a stale tag 412 `etag_mismatch` carrying the current hunt under `current` (and its `ETag` header).",
        "operationId": "updateHunt",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "hunt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hunt Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The hunt's current ETag, e.g. `\"v3\"`.",
              "examples": [
                "\"v3\""
              ],
              "title": "If-Match"
            },
            "description": "The hunt's current ETag, e.g. `\"v3\"`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HuntPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hunt"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "412": {
            "description": "Precondition failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Semantic validation error",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorBody"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Updatehunt"
                }
              }
            }
          },
          "428": {
            "description": "Precondition required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "hunts"
        ],
        "summary": "Delete a saved hunt",
        "description": "`If-Match` is optional here, but it is checked when you send it.",
        "operationId": "deleteHunt",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "hunt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hunt Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The hunt's current ETag, e.g. `\"v3\"`.",
              "examples": [
                "\"v3\""
              ],
              "title": "If-Match"
            },
            "description": "The hunt's current ETag, e.g. `\"v3\"`."
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "412": {
            "description": "Precondition failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable message."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable snake_case error code."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorBody",
        "description": "Error envelope of every domain error. Extra keys carry machine-readable context."
      },
...
      "FilterAll": {
        "properties": {
          "all": {
            "items": {
              "$ref": "#/components/schemas/FilterNode"
            },
            "type": "array",
            "maxItems": 64,
            "title": "All"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "all"
        ],
        "title": "FilterAll"
      },
      "FilterAny": {
        "properties": {
          "any": {
            "items": {
              "$ref": "#/components/schemas/FilterNode"
            },
            "type": "array",
            "maxItems": 64,
            "title": "Any"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "any"
        ],
        "title": "FilterAny"
      },
      "FilterCond": {
        "properties": {
          "field": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Field"
          },
          "op": {
            "$ref": "#/components/schemas/FilterOp"
          },
          "value": {
            "$ref": "#/components/schemas/Scalar"
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/Scalar"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Values"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "op"
        ],
        "title": "FilterCond",
        "description": "A leaf condition. Arity: in -> values (1..50); between -> exactly 2 values; exists -> neither; every other op -> value."
      },
      "FilterNode": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/FilterAll"
          },
          {
            "$ref": "#/components/schemas/FilterAny"
          },
          {
            "$ref": "#/components/schemas/FilterNot"
          },
          {
            "$ref": "#/components/schemas/FilterCond"
          }
        ]
      },
      "FilterNot": {
        "properties": {
          "not": {
            "$ref": "#/components/schemas/FilterNode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "not"
        ],
        "title": "FilterNot"
      },
      "FilterOp": {
        "type": "string",
        "enum": [
          "eq",
          "in",
          "cidr",
          "glob",
          "gte",
          "lte",
          "between",
          "exists"
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Hunt": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "query": {
            "$ref": "#/components/schemas/HuntQuery"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "examples": [
              "2025-10-27T09:14:03.120Z"
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "examples": [
              "2025-10-27T09:14:03.120Z"
            ]
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "name",
          "query",
          "created_at",
          "updated_at",
          "version"
        ],
        "title": "Hunt"
      },
      "HuntCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "title": "Description"
          },
          "query": {
            "$ref": "#/components/schemas/HuntQuery"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "query"
        ],
        "title": "HuntCreate"
      },
      "HuntList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Hunt"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "HuntList"
      },
      "HuntPatch": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "query": {
            "$ref": "#/components/schemas/HuntQuery"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "HuntPatch",
        "description": "JSON merge-patch: omitted keys are unchanged; description: null removes it."
      },
      "HuntPreset": {
        "type": "string",
        "enum": [
          "1h",
          "6h",
          "24h",
          "72h"
        ]
      },
      "HuntQuery": {
        "properties": {
          "sensor_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Sensor Ids"
          },
          "window": {
            "$ref": "#/components/schemas/HuntWindow"
          },
          "filter": {
            "$ref": "#/components/schemas/FilterNode"
          },
          "sort": {
            "$ref": "#/components/schemas/SortKey"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sensor_ids",
          "window",
          "filter",
          "sort"
        ],
        "title": "HuntQuery"
      },
      "HuntWindow": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/HuntWindowPreset"
          },
          {
            "$ref": "#/components/schemas/HuntWindowRange"
          }
        ]
      },
      "HuntWindowPreset": {
        "properties": {
          "preset": {
            "$ref": "#/components/schemas/HuntPreset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "preset"
        ],
        "title": "HuntWindowPreset"
      },
      "HuntWindowRange": {
        "properties": {
          "from": {
            "type": "string",
            "format": "date-time",
            "title": "From",
            "examples": [
              "2025-10-27T09:14:03.120Z"
            ]
          },
          "to": {
            "type": "string",
            "format": "date-time",
            "title": "To",
            "examples": [
              "2025-10-27T09:14:03.120Z"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "from",
          "to"
        ],
        "title": "HuntWindowRange"
      },
...
      "Scalar": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          }
        ]
      },
...
      "SortKey": {
        "type": "string",
        "enum": [
          "ts",
          "-ts",
          "bytes",
          "-bytes",
          "risk",
          "-risk"
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationErrorBody": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ValidationErrorBody",
        "description": "FastAPI's default 422 body: detail is a LIST here, unlike ErrorBody."
      },
      "ValidationIssue": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "ctx": {
            "additionalProperties": true,
            "type": "object",
            "title": "Ctx"
          },
          "input": {
            "title": "Input"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationIssue",
        "description": "One entry of FastAPI's default request-validation error list."
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Access token `at_…` from POST /v1/auth/login or /v1/auth/refresh.",
        "scheme": "bearer"
      }
    }
  },
...
}